backend/services/scheduler/requirements.txt
schedule==1.2.1
requests==2.31.0
python-dotenv==1.0.0
